cast prevent warnings. prepare for dll build

+++ DragonBonesErigon/DragonBonesErigon/src/DragonBones/animation/TimelineState.cpp
@@ -15,6 +15,7 @@
 #include "../armature/Constraint.h"
 #include "../armature/DeformVertices.h"
 #include "AnimationState.h"
+#include <cmath>
 
 DRAGONBONES_NAMESPACE_BEGIN
 
@@ -37,7 +38,7 @@ void ActionTimelineState::_onCrossFrame(unsigned frameIndex) const
             {
                 const auto eventObject = BaseObject::borrowObject<EventObject>();
                 // eventObject->time = _frameArray[frameOffset] * _frameRateR; // Precision problem
-                eventObject->time = _frameArray[frameOffset] / _frameRate;
+                eventObject->time = (float)(_frameArray[frameOffset]) / _frameRate;
                 eventObject->animationState = _animationState;
                 EventObject::actionDataToInstance(action, eventObject, _armature);
                 _armature->_bufferAction(eventObject, true);
@@ -49,7 +50,7 @@ void ActionTimelineState::_onCrossFrame(unsigned frameIndex) const
                 {
                     const auto eventObject = BaseObject::borrowObject<EventObject>();
                     // eventObject->time = _frameArray[frameOffset] * _frameRateR; // Precision problem
-                    eventObject->time = _frameArray[frameOffset] / _frameRate;
+                    eventObject->time = (float)_frameArray[frameOffset] / _frameRate;
                     eventObject->animationState = _animationState;
                     EventObject::actionDataToInstance(action, eventObject, _armature);
                     _armature->_dragonBones->bufferEvent(eventObject);
@@ -667,14 +668,14 @@ void SlotColorTimelineState::_onArriveAtFrame()
     {
         const auto color = slot->_slotData->color;
 
-        _current[0] = color->alphaMultiplier * 100.0f;
-        _current[1] = color->redMultiplier * 100.0f;
-        _current[2] = color->greenMultiplier * 100.0f;
-        _current[3] = color->blueMultiplier * 100.0f;
-        _current[4] = color->alphaOffset;
-        _current[5] = color->redOffset;
-        _current[6] = color->greenOffset;
-        _current[7] = color->blueOffset;
+        _current[0] = static_cast<int>(color->alphaMultiplier * 100.0f);
+        _current[1] = static_cast<int>(color->redMultiplier * 100.0f);
+        _current[2] = static_cast<int>(color->greenMultiplier * 100.0f);
+        _current[3] = static_cast<int>(color->blueMultiplier * 100.0f);
+        _current[4] = static_cast<int>(color->alphaOffset);
+        _current[5] = static_cast<int>(color->redOffset);
+        _current[6] = static_cast<int>(color->greenOffset);
+        _current[7] = static_cast<int>(color->blueOffset);
     }
 }
 
@@ -732,10 +733,10 @@ void SlotColorTimelineState::update(float passedTime)
                 result.redMultiplier += (_result[1] - result.redMultiplier) * fadeProgress;
                 result.greenMultiplier += (_result[2] - result.greenMultiplier) * fadeProgress;
                 result.blueMultiplier += (_result[3] - result.blueMultiplier) * fadeProgress;
-                result.alphaOffset += (_result[4] - result.alphaOffset) * fadeProgress;
-                result.redOffset += (_result[5] - result.redOffset) * fadeProgress;
-                result.greenOffset += (_result[6] - result.greenOffset) * fadeProgress;
-                result.blueOffset += (_result[7] - result.blueOffset) * fadeProgress;
+                result.alphaOffset += static_cast<int>((_result[4] - result.alphaOffset) * fadeProgress);
+                result.redOffset += static_cast<int>((_result[5] - result.redOffset) * fadeProgress);
+                result.greenOffset += static_cast<int>((_result[6] - result.greenOffset) * fadeProgress);
+                result.blueOffset += static_cast<int>((_result[7] - result.blueOffset) * fadeProgress);
 
                 slot->_colorDirty = true;
             }
@@ -758,10 +759,10 @@ void SlotColorTimelineState::update(float passedTime)
                 result.redMultiplier = _result[1];
                 result.greenMultiplier = _result[2];
                 result.blueMultiplier = _result[3];
-                result.alphaOffset = _result[4];
-                result.redOffset = _result[5];
-                result.greenOffset = _result[6];
-                result.blueOffset = _result[7];
+                result.alphaOffset = static_cast<int>(_result[4]);
+                result.redOffset = static_cast<int>(_result[5]);
+                result.greenOffset = static_cast<int>(_result[6]);
+                result.blueOffset = static_cast<int>(_result[7]);
 
                 slot->_colorDirty = true;
             }
@@ -863,7 +864,7 @@ void DeformTimelineState::init(Armature* armature, AnimationState* animationStat
     else 
     {
         const auto deformVertices = slot->_deformVertices;
-        _deformCount = deformVertices != nullptr ? deformVertices->vertices.size() : 0;
+        _deformCount = deformVertices != nullptr ? static_cast<unsigned>(deformVertices->vertices.size()) : 0;
         _valueCount = _deformCount;
         _valueOffset = 0;
         _frameFloatOffset = 0;
@@ -923,7 +924,7 @@ void DeformTimelineState::update(float passedTime)
             }
             else 
             {
-                _deformCount = result.size();
+                _deformCount = static_cast<unsigned>(result.size());
 
                 for (std::size_t i = 0; i < _deformCount; ++i)
                 {
@@ -957,7 +958,7 @@ void DeformTimelineState::update(float passedTime)
             }
             else 
             {
-                _deformCount = result.size();
+                _deformCount = static_cast<unsigned>(result.size());
 
                 for (std::size_t i = 0; i < _deformCount; ++i)
                 {
